Improved form handling of clip areas
Bug #695307 "Preservation of PostScript form elides form content"

When a device (such as pdfwrite) hands back a matrix to set as the CTM
(pdfwrite does this to capture the form unscaled), we need to account for
that CTM when we reset the clip, otherwise we can end up clipping content.

Also, rather than using +/- page boundaries, we use +/- the form BBox.

This shows 2 minor progressions in the test suite.