pdfwrite - consume Distiller params even when LockDistillerparams is true
Bug #695468 "/rangecheck in .installpagedevice error when combining LockDistillerParams, setpagedevice, and restore"

Previously when LockDistillerParams was true the distiller param parsing
code did not get called. This left the Distiller parameter key/value
pairs 'untouched' in the parameter list.

Ordinarily this is not a problem, but .installpagedevice, which is called
during restore to reset the page device parameters, insists that all keys
must be consumed and throws an error if they are not.

This patch alters the distiller parameter parsing to read all the parameters
no matter what the setting of LockDisillerParams. The values are read into
a dummy parameter structure, when complete they are either copied back to
the real structure or, if LockDistillerParams is set, discarded. If we
discard the parameters we clean up the memory rather than relying on garbage
collection to do it for us.

I don't think the pdfwrite device currently does this when the device is
closed, which would be a memory leak, if we ever decide we need to address
this then we can use the code here.

No differences expected.