Fix Bug 695471, cust 532. /SMask /None did not remove prior SMask
The 'gssmask' proc in pdf_draw.ps doesn't remove a previous SMask
when the ExtGState is set to /SMask /None. Change the logic to
issue a "dummy" .begintransparencymaskgroup (that does NOT have a
corresponding .endtransparencymask) and add handling for this type
of SMask in the code. This does not result in a colorspace change,
but only frees the tos maskbuf. Seen with MSN_Chrome.pdf