pdfwrite - mark freshly opened device as 'open'
When using the '%d' format for output files, the pdf_output_page() routine
closes and reopens the pdfwrite device (to flush the device, creating
a PDF file). However, it doesn't mark the 'is_open' member of the device,
and pdfwrite doesn't do that itself, relying on the caller (normally
gs_opendevice()) to set the flag.

Later we get a setpagedevice, this executes .setdevice which calls
gs_setdevice_no_erase() which checks dev->is_open and if it is not open
then it opens it. This causes pdf_open to allocate all kinds of new
memory areas, without freeing the old ones, as well as temporary files and
possibly other structures/memory allocations.

Surprisingly this doesn't actually seem to cause a problem, but its not
right, and surely will cause problems one day, its definitely causing
memory leaks in non-GC interpreters.

So this tiny commit just sets the flag is_open when we open the device in
pdf_output_page.

No differences expected