Free copied descendant font(s) when freeing a copied CIDFont
When we copy a CIDFont (other than a TrueType CIDFont) we also copy the
descendant fonts. But when we free them (gs_free_copied_font) we only free
the CIDFont, leaving the descendants dangling. This can cause problems
(seg fault) when we garbage collect the descendants.

Here we free the descendants as well. This is complicated by the fact that
the parent CIDFont 'glyphs' space is shared by all the descendants, so
we *don't* want to free that when we free the descendant.

No differences expected