pdfwrite - close dangling substreams at end of page
Inspired by bug #695482 but not intended as a fix, and also by bug #695354

If we fail to close a substream (possibly because the PDF interpreter throws
an error) then this leaves a text_state associated with the substream
dangling, and potentially referencing other freed memory which can lead
to a crash in the garbage collector. In addition we can end up writing an
empty PDF file.

This commit closes any substreams which are still open when a page is written
which ensures the memory is cleaned up, and that a valid (if broken) PDF
file gets written.

Note that ps2write is (of course) different, and *expects* to have a substream
open at the end of the page, so in this case we have to close the page and then
clean up any dangling substreams afterwards.

Although this appears to 'fix' bug #695482 its not a proper fix.

No differences expected.