Bug 695502:  segfault with -dBGPrint=true
When we're doing background printing, and we're using a memory file based
clist, we create a "clone" of the file object with it's own "context" but
which shares the underlying "raw" data with the original file. The original
code tries to free the underlying raw data memory along with the clone
file object, but as the clone uses the memory allocator created for the
rendering thread and the raw data memory was allocated in the interpretation
"thread", we end up trying to free memory to the wrong allocator.

This commit has the interpreter thread hold a reference to the original clist
files, so we can free the "clone" files with the thread allocator, and the
original file with the allocator from the interpreter thread. As both these
cleanup actions happen in the interpreter thread, we don't have to worry about
thread clashes.

No cluster differences.