Fix Bug #695535 Page not cleared upon resolution setting.
Somewhat puzzling this hasn't come up before but PCL was changing the
resolution of a PCL job upon receiving a PJL command without erasing
the page.  If the resolution is set on the command line or the device
default is used the normal page erasing happens with the job's
resolution already set.