Bug 695539: glyph mapping - handle "missing" codepoints
When extending the glyph mapping in a font, the original code for adding the
codepoint to the font could assume the codepoint was valid as it was always
read from the Adobe Glyph List. But I added code to "parse" a codepoint from
the special "uniXXXX" type names, and that breaks that assumption.

Handle the case where we reach that point with codepoint that doesn't exist in
the AGL.

No cluster differences.