Make DCT filter finalize routine robust against double calling.
Bug #695542 "Crash on EPS on Mac and Linux"

The problem is that we close a DCT filter inside a gsave context, while
the file ref is still pointing at it.

When the PostScript 'close' operator closes the DCT filter, we free the
memory associated with the stream state in the finalize routine. But the
PostScript file ref is still pointing at it, and when we grestore, the
finalize routine gets called again. Because the routine doesn't check for
NULL pointers before freeing, and indeed doesn't NULL the pointers after it
does free them, this leads to us freeing memory which has already been
freed.

Apparently we've already fixed a number of similar problems in finalize
routines, probably caused by people not realising a finalize routine could
(surprisingly) be called more than once.

In this commit we set the pointers to null after feeing the memory, and
check for null pointers before attempting to free the memory, and before
attempting to dereference any pointers.

No differences expected.