Bug 695244: ensure a consistent allocator for DCT streams.
The existing code could result in a mis-match between the memory allocators
used for the DCT specific parts of the stream and the main body of the stream.

As both are in gc memory, this could result, when the stream goes out of scope,
in the DCT specific memory being garbage collected before the main stream
object. As the main stream object contains the "finalize" method, the call to
finalize would result in accessing already freed memory.

This commit revises the code to use the same method to work out the allocator
to use for the main stream object as we do for the DCT specific parts.

No cluster differences.