pdfwrite - Correct Gray and ColorImageAutoFilterStrategy
A number of problems here; firstly the commit that added these to the
distiller params has the correct names in the commit message, but the names
used in the code were incorrect (eg ColorAutoFilterStrategy).

Secondly the parameter is supposed to be a name, but the data stored
was a C string, and get_params was returning an integer, so it actually
contained half the memory pointer value.

The parameters weren't added to .distillerparamkeys so currentdistillerparams
didn't return the values.

The default value for the non-Luratech case should have been /JPEG but in
fact /JPEG2000 was used in both cases.

99-01-fixed.ps should dhow up as a diff this time, and then should cease
being indeterminate with the pdfwrite device, as we now return the correct
name rather than half a random memory location.