PDF interpreter - improve error recovery with PDF form execution
Bug #695537 "Ghostscript locks up reading attached PDF file"

The problem is that, after we execute a form, we attempt to 'grestore' back
to the gsave depth at the time we executed the form, this is to try and
cater for broken form streams that have unbalanced Q/q operators.

This is done by counting the dict stack before the form execution, and after
the form is complete. If the depth is different we execute 'Q' in a loop
until we get back to the same dict stack depth.

This assumes that the only reason for having more dictionaries on the stack
is due to them being pushed by our implementation of 'q'. However in this
case (and more generally, in any event) we have created some dictionaries
then encountered an error, and returned control back to the form execution
routine.

As stated this calls 'Q' in a loop expecting that each 'Q' will decrement
the number of dictionaries by 1 each time. But 'Q' only does that if the
dictionary is the result of a 'q' if it isn't it leaves it in place.

So the form execution simply carries on round the loop forever.

This commit addresses this by checking the number of items on the dict
stack before and after executing 'Q' and if they are the same, it executes
the 'end' operator instead to remove the topmost dictionary from the stack.

This allows us to carry on removing gstate or general dictionaries until
we get back to the correct depth.

Note that, because there is a genuine error in the file, we do not render
it 'correctly', but neither does any other PDF consumer. Other consumers
render more of it than we do, but they all drop the 'Fannie Mae' logo.

No differences expected