Bug 695146: patch for missing lrintf
OpenJPEG uses lrintf() which can be missing, or in this case, the prototype is
missing from the math.h header file - causing faulty output.

This patch checks for the function prototype, if it's missing it replaces the
lrintf() call with a naive round/cast implementation.

Two reasons for choosing this approach: it saves maintaining a patch against
OpenJPEG, and secondly, this is only relevant to older systems - more up to
date systems have lrintf() correctly prototyped, and often have implementations
optimised to the hardware, so we really want to use it if it's available.

No cluster differences.