This is supposed to be an experiment in implementing FirstPage and LastPage in the device.
The idea was that we would install a forwarding device into the pipeline
when FirstPage/LastPage is set, and the device would consume all operations
on pages which were not between FirstPage and LastPage.

Sadly I can't make this work, there are two effectively insurmountable
problems.

The first is that we cannot create a device pipeline in the way that we did
with Jaws. We can't insert a device in front of an existing device.

The reason is that the device is stored in the graphics state, so the only
way to alter the currently active device is to alter the device in the
graphics state, this is how the pdf14 compositor gets inserted in front
of whichever device is active. It is created as a forwarding device,
pointing to the current device, and the new device returned to the interpreter.
The operator then sets the new device in the graphics state.

Sadly, we don't have the graphics state available to us during put_params
so we cannot have the device create a new forwarding device and push it
in front of itself.

This code creates a new forwarding device and adds it to the build system,
and moves FirstPage and LastPage into the basic device structure, but I
haven't found any way to turn the existing device into a forwarding
device (and still maintain a copy of the existing device to forward to)
nor to create a new forwarding device.