Addition of device 'subclassing'
This is a new feature in the graphcs library which allows for a chain of
devices to be created and for new devices to be created and added to that
chain, even if the graphcs state is not available.

The work for ths was originally performed on the 'device_subclassing'
branch, and all the history is maikntained in that branch. This single
commit embodies all the work, but none of the history.

The background to the code is documented in ghostpdl/doc/subclass.htm.

In addition to the graphics library support for this feature there are two
new devices; the First Page/Last Page device and the Object Filtering
device.