Xpswrite - properly initialise PrinterName
Bug #695617 "Error: Printer name too long"

The xpswrite device doesn't include a full definition of its devicestructure
or initialisers for its variables.

This can lead to problems if any of these variables are exposed via
put/get params as the initial parameter will be uninitialised (put and get
params are called before the device open method)

Here we add a 'finish_copydevice' method and initialise the PrinterName
in it. Devices are created by calling copydevice on the device prototype
and the finish_copydevice method is called as the last action from there
so its safe to initialise the variables, and this will be completed before
put/get params is called.

No differences expected