Bug 695598: cidfmap validation problems on Windows
Originally, the cidfmap validation used "stat" to check whether the font file
was available (and remove the mapping if the font file is missing). But it
seems, for some (currently unidentified) reason, stat can fail even if the
font file is actually avaiable.

This changes the validation code to use ".libfile" and "file" operators to
check whether the font file can be found - this, it seems, works, even in cases
where "state" fails.

No cluster differences.