Add new device for FirstPage and LastPage processing
This temporarily includes some structures and procedures which need to
be moved to the library 'device' processing modules. In here temporarily.

Also adds some new support for 'subclassing' to the device structure, the
macro used to initialise the device structure, and to gdevprn.c to actually
perform the subclassing. This needs (eventually) to be added to gdevvec.c as
well.

Finally, modified the 'bit' device because it also doesn't use the macro
to initialise itself, so we need to update it manually.