PDF interpreter - ignore /Contents array entries which don't resolve to dicts
Bug #695645 "PDFWrite fails on file with Error: /typecheck in --run--"

The specimen file has a /Contents array where one of the elements is an
indirect reference to another array. This isn't legal, entries in the
/Contents array must be references to streams. Acrobat simply ignores the
error and displays the page blank.

Here we check each entry before we process it, and if its not a dictionary
we ignore it, exit the loop processing the entries (as Acrobat apparently does)
and emit a warning.

No differences expected.