Fix annoying, hard to reproduce ICC related warning
Occasionally, and in hard to reproduce circumstances, we could get a warning
from the ICC code:
./base/gsicc_manage.c:1084: gsicc_open_search(): \
Could not find <garbage string>

This, it turns out, is because the get_params code was returning a pointer to
a string allocated in non-gc memory, which just isn't compatible with the
way the Postscript interpreter interacts with the graphics library.

This commit changes the way the parameter is returned so it copies the string
into gc memory, compatible with use in a Postscript string object. Note that the
put_params equivalent code was already correct in copying the string from gc
memory into non-gc memory managed by the device code.

No cluster differences