More work on the garbage collector descriptors for device_printer.
When we enumerate the pointers in the child device, we want to do so with
the actual child device pointer, not the parent. The usage of the pointer
is hidden in the ENUM_PTRS_WITH and ENUM_PREFIX macros, it uses 'vptr' as
the device pointer to pass to the enumeration routine, so we need to set
vptr to the child device before we call it.