More work on GC enumeration. We need to have the GC call the new parent device's enumerator, which I thought would happen by replacing the 'stype' member of the structure. Sadly this is insufficient.
When the structure is created the 'type' member get a pstype vaslue which
is a copy of the stype pointer. I had hoped to just replace the members of
that structure, but it turns out that in general stype_is_dynamic is not
true, and the stype structure is statically defined in the executable which
means we can't overwrite it.

So some ugliness, when subclassing a device we track back to the saved pstype
in the wrapper, and replace that with our own one. That means that the GC
will call our enumerator. From our enumerator we ensure that we call the
child device enumerator (which requires more weirdness due to the macros).

When finalizing the device, we undo all this messing about, at some point
this needs to become an 'unsubclass_device' call.