PDF interpreter - attempt to recover from empty stream with broken endstream
Bug #695664 "Problem reading PDF file: stream length incorrect"

The file has an annotation with a broken stream, its empty (Length 0) and
the endstream keyword is incorrectly written as endsteam.

When we encounter the endobj we attempt to unwind back to the object number
and generation on the stack, but we don't have those on the stack, so instead
we consume the error recovery information, which then leads to the stream
code being unable to recover from the error condition, resulting in an error.

This change to endobj looks to see if either of the 2nd and third entries
on the operand stack are a 'mark', and if so assumes that this is a broken
empty stream, pops back to the mark and exits leaving cleanup to the
stream code.

Clearly this code can be defeated, depending on what garbage is left behind
by the broken stream, but in general we might reasonably expect to find two
integers before a mark object. Even if we don't, the result will still be no
worse than at present.

No differences expected