PDF Interpreter - run annotations in a stopped context for error-and-continue
Bug #695453 "Conversion to png fails with undefinedresult in --run--"

The PDF test file was supplied privately, it contains 3 annotations, each of
which is invalid having /Rect, /QuadPoints or /BBox arrays where each
element is 0.

Here we simply run the annotation inside a stopped context so that if it
throws an error we can raise a message and carry on.

No differences expected.