eps2write - don't clamp text BBox to the text enumerator origin
Bug #695671 "eps2write gives incorrect bounding box with rotated text"

The bug here is that, having gone to some effort to retrieve the bounding
box of the text, the code then clamps the origin of the text box to the
origin of the text enumerator.

This is clearly incorrect for rotated text, and I'm doubtful it makes sense
even for upright text. I'm somewhat puzzled because it looks quite deliberate
but I don't see why, and my testing shows no differences if I don't do this.

So I'm removing this for now, this may need some more attention in the future
but then that is still generally true of the eps2write device.

No differences expected, the eps2write device is not cluster tested.