Fix 695673: symbol set not used correctly.
If we don't support a requested symbol set we map the font directly
with the character codes, i.e. the symbol mapping is identity.  Due to
an historical artifact we used to map symbol and wingding fonts
without symbol sets by just adding 0xf000 to the input character code,
thus evolved the conditional to check if the font is internal and the
symbol set is 0 just add 0xf000 because it must be a wingdings or
symbol font.  Of course that is wrong even if very rarely encountered,
and unnecessary because we've added the symbols sets needed to support
wingdings and symbol.