Bug 695688: improve cidfmap record validation.
The previous record validation only checked for valid substitutions when the
substitution referenced a TTF file (and removed the mapping if the TTF file was
not available). It ignored name to name mappings.

This revision adds validation to the name to name mappings, again removing
ones found to be invalid.

No cluster differences.