Ensure there is a child device before copying parameters from it.
After processing put_params in the child, update the parent device from the
child so that the device properly reflects the child setup.

When creating a compositor, if the returned device is not the same as the
child device, return the new device. Otherwise, replace the returned device
with the parent. Otherwise the overprint code will set the device to the
child device, and we don't want that to happen.