Fix BGPrint=true SEGV and slowdown due to commit 9108db2
Bug 695669:
The BGPrint problem was due to using the fname afer freeing it but the
'fake_path_to_file' logic needed the fname in order to find the real
file to close.

Tail of commit 6ecc8536:
Also the original cfname and bfname (ocfname and obfname) were copied
using strnlen so the trailing NULL wasn't copied resulting in UMR being
used for the close/delete.

Tail of bug 693121:
The performance of unbuffered I/O is slow, seen on many files, but
a good example is: tests/df/Bug6901014_Fourth_crash_signature_testcase.pdf
with params:  -sDEVICE=pbmraw -o /dev/null -r300 -dMaxBitmap=0

The fix for this is to add a read cache. This caching method in intended
to also be used by the memory based clist logic (TBI).

With the cache, the performance on the above goes from 40 seconds down to
4.6 seconds, and with -dBGPrint=true reduces to 3.8 seconds (which crashes
prior to this fix).