We need to create a 'copy_planes' method regardless, this is yet another device method which can be called without checking to see if its NULL, and which is normally set up by the 'open'.
Added some return types to functions and some other tidying up to avoid
compiler warnings.

*MUCH* horrible hackery to work around the fact that the clist
create_compositor does nasty stuff (as it does so much), squirrelling away
copies of some information, replacing it when the device is popped (except
that it isn't really popped, just deactivated) and other stuff. To add to
the unpleasantness, the non-clist create compositor doesn't do any of this
and so we have to differentiate between a clist compositor and a default
compositor. There is no easy way to do this, the pct->type field looked
promising, as there is a 'no_clist_writer' type, but that just means that
its not a clist *writer*, not that its not a clist.... SO I've chosen to
explicitly check the device name, which is just awful. The alternative
(checking method names) is no better IMO.