Provide method for interpreter(s) to add io devices
Previously IO devices were setup in the graphics library at compile/link time.
This adds a method for interpreters to add their own.

Three further enhancements may be required/preferred: this version limits the
number of "dynamic IO devices" to 16 - it would be preferable if the number
were unlimited. Secondly, it would be useful to allow IO devices to be removed/
"unregistered" from the table. Lastly, furthermore it would be good (possibly
necessary) to allow *replacing* of IO devices already "registered"

NOTE: this commit will not work correctly on its own - it needs the following
commit with makefile changes.