Split graphics lib and PS interp configuration...
...into separate headers - the existing gconfig.h (graphics lib) and a new
iconfig.h for the PS interpreter.

This is a big step towards a graphics lib that is fully independendant of the
interpreter.

This also requires that the linker scripts (*.tr files) made product specific,
this is done for gs, but pcl and xps are yet to be done.

Ghostscript (at least) does not work correctly when using the init files from
the romfs - it throws an error a considerable way through initialisation (but
runs when reading the files from disk).