Catering for the case where we need to create a compositor and rendering device for ps2write (or pdfwrite with COmpatibility < 1.4).
When a compositor has been pushed (*pcdev is not our child device) we check to
see if the forwarding device which has been pushed is pointing to our child
device. If it is, then we check to see if its a clist device (by testing
the device name, ick) and proceed to patch up the device appropriately.

If it isn't, then we assume that the compositor has pushed both a compositor
and a rendering device. In this case we need take no action. When rendering
is complete the rendering device will send an image to the target device.