Bug 695711: Resolve issue with relative paths for FONTPATH
Check whether a "lib" file exists, and whether we are allowed to open it,
before we try to open it.

This reverts ""starting_arg_file" should only apply once." (commit
407cc61e87b0fd9d44d72ca740af7d3c85dee78d).

Which caused problems with, for example, the "-sFONTPATH=..." option when the
path is relative rather than absolute.

No cluster differences.