Bug 694509: Fix file name "building" for resources
After we'd created a candidate file name, there was a spurious "exch" in
place before we actually tried to validate the file name.

So, instead of "<path-to-resources>/<Resource Category>/<Resource Instance>"
being passed to .libfile, we actually passed the Postscript "path":
"<Resource Category>/<Resource Instance>"

Simply removing the exch seems to resolve this bug for me.

No cluster differences.