Handle a 'q' during the course of a 'W.....n' sequence
Bug #695723 "Inverted images when converting PDF to Tiff (B&W)"

The 'W' (and W*) operators are unlike their PostScript equivalents, and all
other PDF operators, they don't take effect immediately they simply add to
the clip path and produce a new clip path. Not until the 'n' is executed do
they actually apply the clip path. The Ghostscript PDF interpreter handles
this by accumulating the path operations (W/W*) until a 'n' is encountered.

Normally nothing else would need to be done, but because we've seen broken
PDF files that change the graphics state while accumulating a clip, and we
know that Acrobat ignores these, we have to track this. We do this by using
a dictionary.

So each graphics state operator checks the current dictionary to see if we
are inside a 'W' context and if so emits a warning and drops the change.
Again, so far so good.

However, the 'q' starts a new dictionary (pushing the existing 'W' one back
one level), which means that the 'n' operator closes that new dictionary
instead of the one it should close, and that brings the 'W' dictionary back
to the front.

Now when we execute the '1 1 1 rg' it looks at the dictionary and says "aha,
we are in a 'W' state, so ignore this".

This commit has the 'q' operator check to see if we are in a 'W' context. If
we are we copy the context dictionary to the stack, and close it. Then we
start the q 'nodict' dictionary. Finally we reopen the 'W' dictionary from
the stack. Now the 'n' will correctly close the 'W' dictionary, leaving the
'nodict' in its place.

No differences expected.