xpswrite device:  fix for missing images.
Force vector device to do new path as the clip path is the image path.
I had a case where the clip path ids were the same but the CTM was changing
which resulted in subsequent images coming up missing on the page. i.e. only
the first one was shown.