pdfwrite porpagate the imager state through the pdfwrite Pattern code.
Bug #695733 "Error using pattern under odd conditions"

If we set UseCIEColor, so that all colours are converted into CIEBased spaces,
*and* we set CompatibilityLevel < 1.5 so that we require to convert ICCBased
profiles to V2, and then use a Pattern colour space, pdfwrite throws an
error.

This is because we were not propagating the image state through the pattern
code, and the new colour code requires the imager state in order to generate
a V2 ICC profile for inclusion in the PDF file.

Here we simply pass the imager state through for use.

No differences expected