Add support for encoded source images with xpswrite device.  Fix for bug 695698
Also fixes 695754 and 695680.

The xps format does not allow for the image data encodings that PDF source files can
create.  Hence we have to decode the data prior to writing out to our TIFF file.
The clist image code had static methods for doing the decode during color monitoring.  These
methods were re-factored to work on a new structure that are now member variables for
the xpswrite image enum structure and the clist image enum structure.  The methods and
the structure definition were placed in their own file gximdecode.c/h

Also this adds support for images in indexed, CIELAB, DeviceN and separation color spaces.

Finally, images are stored in a temp file during creation and then copied into the zip
archived, releasing the image temp file.   ICC profiles are written directly from memory
to the zip file.  This avoids having lots of temp files around for each image that is
created.  A similar process needs to be done for other parts of the xps document.  With the
current architecture of having a temp file open for every "file" in the package we would
not be able to create something larger than 512 pages on windows due to the limit on
the number of open files.