Bug 695745: apply resourcestatus hook only once.
pdfwrite requires that Type 32 fonts be disabled, and it does it by hooking
the resourcestatus operator. The problem is, because we always append
to the operator table, if we string a number PS files together from the
PScript5.dll, we can end up overflowing the operator table by repeatedly
doing the resourcestatus hook.

This just puts a boolean into systemdict to note that we're already applied
the hook, and there don't do it again.

No cluster differences.