pdfwrite - some changes to text handling for Tr 3
Bug #695731 "PCL interpreter doesn't send text to the graphics library if its not visible."

Altering the PCL interpreter to set the Text rendering mode to 3 for
'invisible' text demonstrated a few problems in pdfwrite.

1) If the last character drawn was in Tr 3, and we buffer some new text in
Tr mode 0, but we are inside a gsave, and the outer state is Tr 0, then when
we restore back to the Tr 0 the mode is not emitted for the buffered text. Leaving
it as Tr mode 3. To fix this (and potentially other similar graphics state
problems) we now call sync_text_state when we reset text, this will
potentially flush the text buffer as well.

2) We were not using the text rendering mode from the graphics state when
falling back to a type 3 bitmap font, because the character code was out
of range (falling back for other reasons *did* preserve it).

This commit fixes these two problems.

Because of the change to text flushing some text is being drawn 1 pixel different
in location, presumably due to rounding differences, I've reviewed these
and they look fine to me.