Integrate Ken Sharp's solution to preserve PCL transparent text in pdfwrite to fix bug #695731.
PCL characters are rendered transparently when the current color is
white and pattern transparency is on.  Usually this is done by simply
not printing or rendering the effected text.  For high level devices
like pdfwrite we try to preserve the original intent of the source
document and include the text and transparency setting in the output.