pdfwrite - free copied base fonts on exit
Bug #695753 "Segfault in gs_purge_font_from_char_caches_forced with -Z@"

The problem here is that we free copied fonts when we free the associated
FontDescriptor (because until then we need the copied font). However, if
we have a 'base 14' font, or other font which is not embedded, we do not
create a FontDescriptor, and so never free the font copy we take when the
font is used.

This later causes a problem because when the garbage collector frees this
font copy it is still pointing to a font cache which no longer exists.

This commit tests to see if a FontDescriptor is associated with a font when
we free the (PDF) font. If it is not then we free the copied font there and
then, as well as the ancillary bytes used for the font name etc.

No differences expected.