Fix bug 695373: xpswrite leaves temp files on Windows.
There is a recent addidtion of a gp function to open a temp file that
will be deleted upon close: 'gp_open_scratch_file_rm'. The method of
'unlink' after open does not work on Windows. This function is platform
specific and works on all supported platforms even if the process crashes.

Also get rid of confusing error messages. PageSpotColors is now only
given to devices that support it (pdf_main.ps) and NumRenderingThreads
and BGPrint are ignored for vector devices (gdevvec.c).