Bug 695772: account for clist icc_table being shared
When multi-thread rendering is used (but not background printing), the icc_table
is shared between all the render thread devices and the 'parent' device in the
interpreter thread - but it is not reference counted. We rely on the fact that
the device in the interpreter thread will exist before, and persist past the
death of the rendering threads.

So, when dismantling a rendering thread device, null the icc_table entry so
it doesn't get erroneously freed.

No cluster differences.