pdfwrite - optimise pdf_foget_resource with CharProc resources
Bug #695778 "Long delay converting .ps file to .pdf file"

The specimen file uses glyphshow to draw a huge number of 'characters',
these turn out to be bitmaps which, taken together create the page. Each
page defines and at the end specifically undefines a type 3 font and the
glyphs are stored in this font.

Since PDF has no equivalent to glypshow we store the glyphs in multiple
type 3 fonts, resulting in around 10 times as many fonts. The creation and
storage of the fonts and glyphs is quite slow.

However, we start off unaware that this is a glyphshow operation, so we
start by attempting to capture the type 3 CharProc. Its not until we get
to the actual capture of the bitmap that we discover we cannot handle the
glyph like this. That means we have to discard the CharProc we had begun,
and remove it from the list of stored resources.

Now resources are stored in 'chains' which is a simple way of limiting the
number of searches. We hash on the resource ID and pick the chain based on
that.

The first problem was that pdf_forget_resource was not using the hash, it
was checking every entry in every chain, in effect checking all the entries
of a given resource type. Even if it found a match, it continued checking
the remaining chains!  As the number of stored CharProcs (we still store
the glyphs as CharProcs, but in a different font) increases, the time taken
to exhaustively check all the glyphs increases.

Here we fix that by starting from the correct chain based on the hash of
the id (if its 0 we start at the first chain, so if we haven't allocated an
ID we check all the entries). When we find a match we return immediately
instead of merely exiting the *inner* loop.

The second problem is that we don't allocate an ID for CharProcs initially.
This is fixed by storing the ID which we used temporarily as an argument
to pdf_enter_substream.

Taken together these speed up this file by a factor of ~8.

No differences expected.